schema/requirements.txt
jsonschema==3.1.0
pyyaml==5.3.1
